backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic-settings
alembic
python-dotenv
pytest
httpx
PyJWT
passlib[argon2]
